Move asynchronous CIB handling to cib/
